requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
ccxt==4.1.77
pandas==2.1.3
numpy==1.25.2
ta==0.10.2
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.1.0
